refactor: split SinglePost component. Add test-ids Splitted SinglePost into SinglePost, CommentComponent and CommentReply. Added a couple of data-testids needed by playwright tests.
